# exec: add support for RIGHT OUTER in merge join