--- a/pkg/sql/exec/execgen/cmd/execgen/mergejoiner_gen.go
+++ b/pkg/sql/exec/execgen/cmd/execgen/mergejoiner_gen.go
@@ -41,9 +41,11 @@ type selPermutation struct {
 }
 
 type joinTypeInfo struct {
-	IsInner     bool
-	IsLeftOuter bool
-	String      string
+	IsInner      bool
+	IsLeftOuter  bool
+	IsRightOuter bool
+
+	String string
 }
 
 func genMergeJoinOps(wr io.Writer) error {
@@ -72,11 +74,20 @@ func genMergeJoinOps(wr io.Writer) error {
 	leftUnmatchedGroupSwitch := makeFunctionRegex("_LEFT_UNMATCHED_GROUP_SWITCH", 1)
 	s = leftUnmatchedGroupSwitch.ReplaceAllString(s, `{{template "leftUnmatchedGroupSwitch" buildDict "Global" $ "JoinType" $1}}`)
 
+	rightUnmatchedGroupSwitch := makeFunctionRegex("_RIGHT_UNMATCHED_GROUP_SWITCH", 1)
+	s = rightUnmatchedGroupSwitch.ReplaceAllString(s, `{{template "rightUnmatchedGroupSwitch" buildDict "Global" $ "JoinType" $1}}`)
+
 	nullFromLeftSwitch := makeFunctionRegex("_NULL_FROM_LEFT_SWITCH", 1)
 	s = nullFromLeftSwitch.ReplaceAllString(s, `{{template "nullFromLeftSwitch" buildDict "Global" $ "JoinType" $1}}`)
 
+	nullFromRightSwitch := makeFunctionRegex("_NULL_FROM_RIGHT_SWITCH", 1)
+	s = nullFromRightSwitch.ReplaceAllString(s, `{{template "nullFromRightSwitch" buildDict "Global" $ "JoinType" $1}}`)
+
 	incrementLeftSwitch := makeFunctionRegex("_INCREMENT_LEFT_SWITCH", 4)
 	s = incrementLeftSwitch.ReplaceAllString(s, `{{template "incrementLeftSwitch" buildDict "Global" $ "JoinType" $1 "Sel" $2 "MJOverload" $3 "lHasNulls" $4}}`)
+
+	incrementRightSwitch := makeFunctionRegex("_INCREMENT_RIGHT_SWITCH", 4)
+	s = incrementRightSwitch.ReplaceAllString(s, `{{template "incrementRightSwitch" buildDict "Global" $ "JoinType" $1 "Sel" $2 "MJOverload" $3 "RNull" $4}}`)
 
 	processNotLastGroupInColumnSwitch := makeFunctionRegex("_PROCESS_NOT_LAST_GROUP_IN_COLUMN_SWITCH", 1)
 	s = processNotLastGroupInColumnSwitch.ReplaceAllString(s, `{{template "processNotLastGroupInColumnSwitch" buildDict "Global" $ "JoinType" $1}}`)
@@ -92,9 +103,6 @@ func genMergeJoinOps(wr io.Writer) error {
 
 	rightSwitch := makeFunctionRegex("_RIGHT_SWITCH", 2)
 	s = rightSwitch.ReplaceAllString(s, `{{template "rightSwitch" buildDict "Global" $ "IsSel" $1  "HasNulls" $2 }}`)
-
-	nullInBufferedGroupSwitch := makeFunctionRegex("_NULL_IN_BUFFERED_GROUP_SWITCH", 1)
-	s = nullInBufferedGroupSwitch.ReplaceAllString(s, `{{template "nullInBufferedGroupSwitch" buildDict "Global" $ "JoinType" $1}}`)
 
 	assignEqRe := makeFunctionRegex("_ASSIGN_EQ", 3)
 	s = assignEqRe.ReplaceAllString(s, `{{.Eq.Assign $1 $2 $3}}`)
@@ -161,6 +169,10 @@ func genMergeJoinOps(wr io.Writer) error {
 		{
 			IsLeftOuter: true,
 			String:      "LeftOuter",
+		},
+		{
+			IsRightOuter: true,
+			String:       "RightOuter",
 		},
 	}