--- a/pkg/sql/exec/mergejoiner_tmpl.go
+++ b/pkg/sql/exec/mergejoiner_tmpl.go
@@ -118,6 +118,7 @@ func _PROBE_SWITCH(
 			curRLength := rGroup.rowEndIdx
 			areGroupsProcessed := false
 			_LEFT_UNMATCHED_GROUP_SWITCH(_JOIN_TYPE)
+			_RIGHT_UNMATCHED_GROUP_SWITCH(_JOIN_TYPE)
 			// Expand or filter each group based on the current equality column.
 			for curLIdx < curLLength && curRIdx < curRLength && !areGroupsProcessed {
 				// TODO(georgeutsin): change null check logic for non INNER joins.
@@ -130,6 +131,7 @@ func _PROBE_SWITCH(
 				// {{ end }}
 				// {{ if _R_HAS_NULLS }}
 				if rVec.Nulls().NullAt64(uint64(_R_SEL_IND)) {
+					_NULL_FROM_RIGHT_SWITCH(_JOIN_TYPE)
 					curRIdx++
 					continue
 				}
@@ -225,6 +227,11 @@ func _PROBE_SWITCH(
 						// {{ end }}
 					} else {
 						curRIdx++
+						// {{ if $.RNull }}
+						_INCREMENT_RIGHT_SWITCH(_JOIN_TYPE, _SEL_ARG, _MJ_OVERLOAD, true)
+						// {{ else }}
+						_INCREMENT_RIGHT_SWITCH(_JOIN_TYPE, _SEL_ARG, _MJ_OVERLOAD, false)
+						// {{ end }}
 					}
 
 				}
@@ -265,6 +272,43 @@ func _LEFT_UNMATCHED_GROUP_SWITCH(joinType joinTypeInfo) { // */}}
 		curLIdx++
 		areGroupsProcessed = true
 	}
+	// {{ else if $.JoinType.IsRightOuter }}
+	// {{/*
+	// Unmatched groups from the left are not possible with RIGHT OUTER JOIN, so
+	// there is nothing to do here.
+	// */}}
+	// {{ end }}
+	// {{end}}
+	// {{/*
+}
+
+// */}}
+
+// {{/*
+// This code snippet processes an unmatched group from the right.
+func _RIGHT_UNMATCHED_GROUP_SWITCH(joinType joinTypeInfo) { // */}}
+	// {{define "rightUnmatchedGroupSwitch"}}
+	// {{ if $.JoinType.IsInner }}
+	// {{/*
+	// Unmatched groups are not possible with INNER JOIN, so there is nothing to
+	// do here.
+	// */}}
+	// {{ else if $.JoinType.IsLeftOuter }}
+	// {{/*
+	// Unmatched groups from the right are not possible with LEFT OUTER JOIN, so
+	// there is nothing to do here.
+	// */}}
+	// {{ else if $.JoinType.IsRightOuter }}
+	if rGroup.unmatched {
+		if curRIdx+1 != curRLength {
+			panic("unexpectedly length of the right unmatched group is not 1")
+		}
+		// The row already does not have a match, so we don't need to do any
+		// additional processing.
+		o.groups.addRightOuterGroup(curLIdx, curRIdx)
+		curRIdx++
+		areGroupsProcessed = true
+	}
 	// {{ end }}
 	// {{end}}
 	// {{/*
@@ -283,6 +327,32 @@ func _NULL_FROM_LEFT_SWITCH(joinType joinTypeInfo) { // */}}
 	// */}}
 	// {{ else if $.JoinType.IsLeftOuter }}
 	o.groups.addLeftOuterGroup(curLIdx, curRIdx)
+	// {{ else if $.JoinType.IsRightOuter }}
+	// {{/*
+	// Nulls coming from the left input are ignored in RIGHT OUTER JOIN.
+	// */}}
+	// {{ end }}
+	// {{end}}
+	// {{/*
+}
+
+// */}}
+
+// {{/*
+// This code snippet decides what to do if we encounter null in the equality
+// column from the right input.
+func _NULL_FROM_RIGHT_SWITCH(joinType joinTypeInfo) { // */}}
+	// {{define "nullFromRightSwitch"}}
+	// {{ if $.JoinType.IsInner }}
+	// {{/*
+	// Nulls coming from the right input are ignored in INNER JOIN.
+	// */}}
+	// {{ else if $.JoinType.IsLeftOuter }}
+	// {{/*
+	// Nulls coming from the right input are ignored in LEFT OUTER JOIN.
+	// */}}
+	// {{ else if $.JoinType.IsRightOuter }}
+	o.groups.addRightOuterGroup(curLIdx, curRIdx)
 	// {{ end }}
 	// {{end}}
 	// {{/*
@@ -325,6 +395,55 @@ func _INCREMENT_LEFT_SWITCH(
 		o.groups.addLeftOuterGroup(curLIdx, curRIdx)
 		curLIdx++
 	}
+	// {{ else if $.JoinType.IsRightOuter }}
+	// {{/*
+	// Unmatched tuple from the left source is not outputted in RIGHT OUTER JOIN.
+	// */}}
+	// {{ end }}
+	// {{end}}
+	// {{/*
+}
+
+// */}}
+
+// {{/*
+// This code snippet decides what to do when - while looking for a match
+// between two inputs - we need to advance the right side, i.e. it decides how
+// to handle an unmatched tuple from the right.
+func _INCREMENT_RIGHT_SWITCH(
+	joinType joinTypeInfo, sel selPermutation, overload mjOverload, rHasNulls bool,
+) { // */}}
+	// {{define "incrementRightSwitch"}}
+	// {{ $sel := $.Sel }}
+	// {{ $mjOverload := $.MJOverload }}
+	// {{ if $.JoinType.IsInner }}
+	// {{/*
+	// Unmatched tuple from the right source is not outputted in INNER JOIN.
+	// */}}
+	// {{ else if $.JoinType.IsLeftOuter }}
+	// {{/*
+	// Unmatched tuple from the right source is not outputted in LEFT OUTER JOIN.
+	// */}}
+	// {{ else if $.JoinType.IsRightOuter }}
+	// All the rows on the right within the current group will not get a match on
+	// the left, so we're adding each of them as a right outer group.
+	o.groups.addRightOuterGroup(curLIdx, curRIdx-1)
+	for curRIdx < curRLength {
+		// {{ if $.RNull }}
+		if rVec.Nulls().NullAt64(uint64(_R_SEL_IND)) {
+			break
+		}
+		// {{ end }}
+		newRVal := rKeys[_R_SEL_IND]
+		// {{with $mjOverload}}
+		_ASSIGN_EQ("match", "newRVal", "rVal")
+		// {{end}}
+		if !match {
+			break
+		}
+		o.groups.addRightOuterGroup(curLIdx, curRIdx)
+		curRIdx++
+	}
 	// {{ end }}
 	// {{end}}
 	// {{/*
@@ -352,6 +471,19 @@ func _PROCESS_NOT_LAST_GROUP_IN_COLUMN_SWITCH(joinType joinTypeInfo) { // */}}
 		for curLIdx < curLLength {
 			o.groups.addLeftOuterGroup(curLIdx, curRIdx)
 			curLIdx++
+		}
+	}
+	// {{ else if $.JoinType.IsRightOuter }}
+	if !o.groups.isLastGroupInCol() && !areGroupsProcessed {
+		// The current group is not the last one within the column, so it
+		// cannot be extended into the next batch, and we need to process it
+		// right now.
+		// Any unprocessed row in the right group will not get a match, so
+		// each one of them becomes a new unmatched group with a
+		// corresponding null group.
+		for curRIdx < curRLength {
+			o.groups.addRightOuterGroup(curLIdx, curRIdx)
+			curRIdx++
 		}
 	}
 	// {{ end }}
@@ -406,8 +538,8 @@ EqLoop:
 	}
 }
 
-// {{end}}
-// {{end}}
+// {{ end }}
+// {{ end }}
 
 // {{/*
 // This code snippet builds the output corresponding to the left side (i.e. is
@@ -445,20 +577,25 @@ func _LEFT_SWITCH(isSel bool, hasNulls bool) { // */}}
 					toAppend = outputBatchSize - outStartIdx
 				}
 
-				var isNull bool
-				// {{ if $.HasNulls }}
-				isNull = src.Nulls().NullAt64(uint64(srcStartIdx))
-				if isNull {
+				if leftGroup.nullGroup {
 					out.Nulls().SetNullRange(uint64(outStartIdx), uint64(outStartIdx+toAppend))
 					outStartIdx += toAppend
-				}
-				// {{ end }}
+				} else {
+					var isNull bool
+					// {{ if $.HasNulls }}
+					isNull = src.Nulls().NullAt64(uint64(srcStartIdx))
+					if isNull {
+						out.Nulls().SetNullRange(uint64(outStartIdx), uint64(outStartIdx+toAppend))
+						outStartIdx += toAppend
+					}
+					// {{ end }}
 
-				if !isNull {
-					val = srcCol[srcStartIdx]
-					for i := 0; i < toAppend; i++ {
-						outCol[outStartIdx] = val
-						outStartIdx++
+					if !isNull {
+						val = srcCol[srcStartIdx]
+						for i := 0; i < toAppend; i++ {
+							outCol[outStartIdx] = val
+							outStartIdx++
+						}
 					}
 				}
 
@@ -797,6 +934,20 @@ func _SOURCE_FINISHED_SWITCH(joinType joinTypeInfo) { // */}}
 		o.exhaustLeftSourceForLeftOuter()
 		// We do not set outputReady here to true because we want to put as
 		// many unmatched tuples from the left into the output batch. Once
+		// outCount reaches the desired output batch size, the output will be
+		// returned.
+	} else {
+		o.outputReady = true
+	}
+	// {{ else if $.JoinType.IsRightOuter }}
+	// At least one of the sources is finished. If it was the left one,
+	// then we need to emit remaining tuples from the right source with
+	// nulls corresponding to the left one. But if the right source is
+	// finished, then there is nothing left to do.
+	if o.proberState.rIdx < o.proberState.rLength {
+		o.exhaustRightSourceForRightOuter()
+		// We do not set outputReady here to true because we want to put as
+		// many unmatched tuples from the right into the output batch. Once
 		// outCount reaches the desired output batch size, the output will be
 		// returned.
 	} else {