--- a/pkg/sql/exec/mergejoiner_tmpl.go
+++ b/pkg/sql/exec/mergejoiner_tmpl.go
@@ -121,7 +121,6 @@ func _PROBE_SWITCH(
 			_RIGHT_UNMATCHED_GROUP_SWITCH(_JOIN_TYPE)
 			// Expand or filter each group based on the current equality column.
 			for curLIdx < curLLength && curRIdx < curRLength && !areGroupsProcessed {
-				// TODO(georgeutsin): change null check logic for non INNER joins.
 				// {{ if _L_HAS_NULLS }}
 				if lVec.Nulls().NullAt64(uint64(_L_SEL_IND)) {
 					_NULL_FROM_LEFT_SWITCH(_JOIN_TYPE)
@@ -154,7 +153,6 @@ func _PROBE_SWITCH(
 					} else {
 						curLIdx++
 						for curLIdx < curLLength {
-							// TODO(georgeutsin): change null check logic for non INNER joins.
 							// {{ if _L_HAS_NULLS }}
 							if lVec.Nulls().NullAt64(uint64(_L_SEL_IND)) {
 								lComplete = true
@@ -178,7 +176,6 @@ func _PROBE_SWITCH(
 					} else {
 						curRIdx++
 						for curRIdx < curRLength {
-							// TODO(georgeutsin): change null check logic for non INNER joins.
 							// {{ if _R_HAS_NULLS }}
 							if rVec.Nulls().NullAt64(uint64(_R_SEL_IND)) {
 								rComplete = true
@@ -227,7 +224,7 @@ func _PROBE_SWITCH(
 						// {{ end }}
 					} else {
 						curRIdx++
-						// {{ if $.RNull }}
+						// {{ if _R_HAS_NULLS }}
 						_INCREMENT_RIGHT_SWITCH(_JOIN_TYPE, _SEL_ARG, _MJ_OVERLOAD, true)
 						// {{ else }}
 						_INCREMENT_RIGHT_SWITCH(_JOIN_TYPE, _SEL_ARG, _MJ_OVERLOAD, false)
@@ -264,7 +261,7 @@ func _LEFT_UNMATCHED_GROUP_SWITCH(joinType joinTypeInfo) { // */}}
 	// {{ else if $.JoinType.IsLeftOuter }}
 	if lGroup.unmatched {
 		if curLIdx+1 != curLLength {
-			panic("unexpectedly length of the left unmatched group is not 1")
+			panic(fmt.Sprintf("unexpectedly length %d of the left unmatched group is not 1", curLLength-curLIdx))
 		}
 		// The row already does not have a match, so we don't need to do any
 		// additional processing.
@@ -301,7 +298,7 @@ func _RIGHT_UNMATCHED_GROUP_SWITCH(joinType joinTypeInfo) { // */}}
 	// {{ else if $.JoinType.IsRightOuter }}
 	if rGroup.unmatched {
 		if curRIdx+1 != curRLength {
-			panic("unexpectedly length of the right unmatched group is not 1")
+			panic(fmt.Sprintf("unexpectedly length %d of the right unmatched group is not 1", curRLength-curRIdx))
 		}
 		// The row already does not have a match, so we don't need to do any
 		// additional processing.
@@ -429,7 +426,7 @@ func _INCREMENT_RIGHT_SWITCH(
 	// the left, so we're adding each of them as a right outer group.
 	o.groups.addRightOuterGroup(curLIdx, curRIdx-1)
 	for curRIdx < curRLength {
-		// {{ if $.RNull }}
+		// {{ if _R_HAS_NULLS }}
 		if rVec.Nulls().NullAt64(uint64(_R_SEL_IND)) {
 			break
 		}
@@ -834,12 +831,16 @@ func (o *mergeJoinBase) isBufferedGroupFinished(
 	if batch.Length() == 0 {
 		return true
 	}
-	sel := batch.Selection()
 	bufferedGroup := o.proberState.lBufferedGroup
 	if input == &o.right {
 		bufferedGroup = o.proberState.rBufferedGroup
 	}
 	lastBufferedTupleIdx := bufferedGroup.length - 1
+	tupleToLookAtIdx := uint64(rowIdx)
+	sel := batch.Selection()
+	if sel != nil {
+		tupleToLookAtIdx = uint64(sel[rowIdx])
+	}
 
 	// Check all equality columns in the first row of batch to make sure we're in
 	// the same group.
@@ -859,19 +860,10 @@ func (o *mergeJoinBase) isBufferedGroupFinished(
 			}
 			prevVal := bufferedGroup.ColVec(int(colIdx))._TemplateType()[lastBufferedTupleIdx]
 			var curVal _GOTYPE
-			if sel != nil {
-				// TODO (georgeutsin): Potentially update this logic for non INNER joins.
-				if batch.ColVec(int(colIdx)).MaybeHasNulls() && batch.ColVec(int(colIdx)).Nulls().NullAt64(uint64(sel[rowIdx])) {
-					return true
-				}
-				curVal = batch.ColVec(int(colIdx))._TemplateType()[sel[rowIdx]]
-			} else {
-				// TODO (georgeutsin): Potentially update this logic for non INNER joins.
-				if batch.ColVec(int(colIdx)).MaybeHasNulls() && batch.ColVec(int(colIdx)).Nulls().NullAt64(uint64(rowIdx)) {
-					return true
-				}
-				curVal = batch.ColVec(int(colIdx))._TemplateType()[rowIdx]
+			if batch.ColVec(int(colIdx)).MaybeHasNulls() && batch.ColVec(int(colIdx)).Nulls().NullAt64(tupleToLookAtIdx) {
+				return true
 			}
+			curVal = batch.ColVec(int(colIdx))._TemplateType()[tupleToLookAtIdx]
 			var match bool
 			_ASSIGN_EQ("match", "prevVal", "curVal")
 			if !match {
@@ -916,6 +908,94 @@ func (o *mergeJoin_JOIN_TYPE_STRINGOp) probe() {
 
 // {{ end }}
 
+// {{ range $joinType := .JoinTypes }}
+
+// exhaustLeftSource sets up the builder to process any remaining tuples from
+// the left source. It should only be called when the right source has been
+// exhausted.
+func (o *mergeJoin_JOIN_TYPE_STRINGOp) exhaustLeftSource() {
+	// {{ if $joinType.IsInner }}
+	// {{/*
+	// Remaining tuples from the left source do not have a match, so they are
+	// ignored in INNER JOIN.
+	// */}}
+	// {{ else if $joinType.IsLeftOuter }}
+	// The capacity of builder state lGroups and rGroups is always at least 1
+	// given the init.
+	o.builderState.lGroups = o.builderState.lGroups[:1]
+	o.builderState.lGroups[0] = group{
+		rowStartIdx: o.proberState.lIdx,
+		rowEndIdx:   o.proberState.lLength,
+		numRepeats:  1,
+		toBuild:     o.proberState.lLength - o.proberState.lIdx,
+		unmatched:   true,
+	}
+	o.builderState.rGroups = o.builderState.rGroups[:1]
+	o.builderState.rGroups[0] = group{
+		rowStartIdx: o.proberState.lIdx,
+		rowEndIdx:   o.proberState.lLength,
+		numRepeats:  1,
+		toBuild:     o.proberState.lLength - o.proberState.lIdx,
+		nullGroup:   true,
+	}
+	o.builderState.lBatch = o.proberState.lBatch
+	o.builderState.rBatch = o.proberState.rBatch
+
+	o.proberState.lIdx = o.proberState.lLength
+	// {{ else if $joinType.IsRightOuter }}
+	// {{/*
+	// Remaining tuples from the left source do not have a match, so they are
+	// ignored in RIGHT OUTER JOIN.
+	// */}}
+	// {{ end }}
+}
+
+// {{ end }}
+
+// {{ range $joinType := .JoinTypes }}
+
+// exhaustRightSource sets up the builder to process any remaining tuples from
+// the right source. It should only be called when the left source has been
+// exhausted.
+func (o *mergeJoin_JOIN_TYPE_STRINGOp) exhaustRightSource() {
+	// {{ if $joinType.IsInner }}
+	// {{/*
+	// Remaining tuples from the right source do not have a match, so they are
+	// ignored in INNER JOIN.
+	// */}}
+	// {{ else if $joinType.IsLeftOuter }}
+	// {{/*
+	// Remaining tuples from the right source do not have a match, so they are
+	// ignored in LEFT OUTER JOIN.
+	// */}}
+	// {{ else if $joinType.IsRightOuter }}
+	// The capacity of builder state lGroups and rGroups is always at least 1
+	// given the init.
+	o.builderState.lGroups = o.builderState.lGroups[:1]
+	o.builderState.lGroups[0] = group{
+		rowStartIdx: o.proberState.rIdx,
+		rowEndIdx:   o.proberState.rLength,
+		numRepeats:  1,
+		toBuild:     o.proberState.rLength - o.proberState.rIdx,
+		nullGroup:   true,
+	}
+	o.builderState.rGroups = o.builderState.rGroups[:1]
+	o.builderState.rGroups[0] = group{
+		rowStartIdx: o.proberState.rIdx,
+		rowEndIdx:   o.proberState.rLength,
+		numRepeats:  1,
+		toBuild:     o.proberState.rLength - o.proberState.rIdx,
+		unmatched:   true,
+	}
+	o.builderState.lBatch = o.proberState.lBatch
+	o.builderState.rBatch = o.proberState.rBatch
+
+	o.proberState.rIdx = o.proberState.rLength
+	// {{ end }}
+}
+
+// {{ end }}
+
 // {{/*
 // This code snippet is executed when at least one of the input sources has
 // been exhausted. It processes any remaining tuples and then sets up the
@@ -931,7 +1011,7 @@ func _SOURCE_FINISHED_SWITCH(joinType joinTypeInfo) { // */}}
 	// nulls corresponding to the right one. But if the left source is
 	// finished, then there is nothing left to do.
 	if o.proberState.lIdx < o.proberState.lLength {
-		o.exhaustLeftSourceForLeftOuter()
+		o.exhaustLeftSource()
 		// We do not set outputReady here to true because we want to put as
 		// many unmatched tuples from the left into the output batch. Once
 		// outCount reaches the desired output batch size, the output will be
@@ -945,7 +1025,7 @@ func _SOURCE_FINISHED_SWITCH(joinType joinTypeInfo) { // */}}
 	// nulls corresponding to the left one. But if the right source is
 	// finished, then there is nothing left to do.
 	if o.proberState.rIdx < o.proberState.rLength {
-		o.exhaustRightSourceForRightOuter()
+		o.exhaustRightSource()
 		// We do not set outputReady here to true because we want to put as
 		// many unmatched tuples from the right into the output batch. Once
 		// outCount reaches the desired output batch size, the output will be