… pkg/sql/distsqlrun/columnar_operators_test.go
… pkg/sql/exec/mergejoiner_test.go