First commit adds support for RIGHT OUTER JOIN in vectorized merge join.

Second commits does some housekeeping.

Addresses: #37166.