@yuzefovich 2019-07-02
An interesting observation is that increasing the size of the generated `mergejoiner.eg.go` (roughly from 80k lines to 120k lines) gives a very minor performance hit (note that the benchmarks are for INNER JOIN, and the generated code for that should be exactly the same):
```
name                                       old time/op    new time/op    delta
MergeJoiner/rows=1024-24                     51.5µs ± 2%    51.6µs ± 3%    ~     (p=0.971 n=10+10)
MergeJoiner/rows=4096-24                      177µs ± 1%     179µs ± 2%  +1.39%  (p=0.003 n=10+10)
MergeJoiner/rows=16384-24                     661µs ± 1%     671µs ± 1%  +1.47%  (p=0.000 n=10+9)
MergeJoiner/rows=1048576-24                  41.4ms ± 1%    42.1ms ± 2%  +1.63%  (p=0.000 n=10+10)
MergeJoiner/oneSideRepeat-rows=1024-24       54.0µs ± 1%    54.1µs ± 1%    ~     (p=0.387 n=9+9)
MergeJoiner/oneSideRepeat-rows=4096-24        180µs ± 1%     181µs ± 1%    ~     (p=0.400 n=10+9)
MergeJoiner/oneSideRepeat-rows=16384-24       673µs ± 1%     673µs ± 1%    ~     (p=0.796 n=10+10)
MergeJoiner/oneSideRepeat-rows=1048576-24    41.7ms ± 1%    42.1ms ± 1%  +1.09%  (p=0.000 n=10+10)
MergeJoiner/bothSidesRepeat-rows=1024-24     56.0µs ± 2%    57.1µs ± 1%  +1.90%  (p=0.000 n=10+10)
MergeJoiner/bothSidesRepeat-rows=4096-24      200µs ± 2%     197µs ± 1%  -1.24%  (p=0.005 n=10+10)
MergeJoiner/bothSidesRepeat-rows=16384-24    1.06ms ± 0%    1.07ms ± 0%  +0.43%  (p=0.000 n=10+10)
MergeJoiner/bothSidesRepeat-rows=32768-24    3.11ms ± 0%    3.14ms ± 0%  +0.77%  (p=0.000 n=9+9)

name                                       old speed      new speed      delta
MergeJoiner/rows=1024-24                   1.27GB/s ± 2%  1.27GB/s ± 3%    ~     (p=0.971 n=10+10)
MergeJoiner/rows=4096-24                   1.48GB/s ± 1%  1.46GB/s ± 1%  -1.59%  (p=0.000 n=10+9)
MergeJoiner/rows=16384-24                  1.59GB/s ± 1%  1.56GB/s ± 1%  -1.45%  (p=0.000 n=10+9)
MergeJoiner/rows=1048576-24                1.62GB/s ± 1%  1.59GB/s ± 2%  -1.60%  (p=0.000 n=10+10)
MergeJoiner/oneSideRepeat-rows=1024-24     1.21GB/s ± 1%  1.21GB/s ± 1%    ~     (p=0.387 n=9+9)
MergeJoiner/oneSideRepeat-rows=4096-24     1.45GB/s ± 1%  1.45GB/s ± 1%    ~     (p=0.400 n=10+9)
MergeJoiner/oneSideRepeat-rows=16384-24    1.56GB/s ± 1%  1.56GB/s ± 1%    ~     (p=0.796 n=10+10)
MergeJoiner/oneSideRepeat-rows=1048576-24  1.61GB/s ± 1%  1.59GB/s ± 1%  -1.07%  (p=0.000 n=10+10)
MergeJoiner/bothSidesRepeat-rows=1024-24   1.17GB/s ± 2%  1.15GB/s ± 1%  -1.87%  (p=0.000 n=10+10)
MergeJoiner/bothSidesRepeat-rows=4096-24   1.31GB/s ± 2%  1.33GB/s ± 1%  +1.24%  (p=0.005 n=10+10)
MergeJoiner/bothSidesRepeat-rows=16384-24   987MB/s ± 0%   983MB/s ± 0%  -0.43%  (p=0.000 n=10+10)
MergeJoiner/bothSidesRepeat-rows=32768-24   674MB/s ± 0%   668MB/s ± 0%  -0.76%  (p=0.000 n=9+9)
```