@yuzefovich 2019-07-02
I've just added another commit to this PR that cleans up the code a bit (removes the TODOs, runs tests against the processor several time - locally takes 0.4s, can reduce it further if needed, renames a couple of methods). If you think it doesn't belong here, please let me know, and I'll move it into the follow-up PR to this one.