@yuzefovich 2019-07-08
Rebased it on top of master, so it is RFAL.